requirements.txt
azure-core==1.26.1
azure-storage-blob==12.14.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
et-xmlfile==1.1.0
Flask==2.2.2
idna==3.4
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
minidom-ext==0.0.5
msrest==0.7.1
ntlm-auth==1.5.0
oauthlib==3.2.2
openpyxl==3.0.10
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
SharePlum==0.5.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
